plugins-details-CTA, -sidebar: extract logic to new function usePluginsSupportText
